+++ packages/build-tools/cocoapods-plugin/lib/plugin.rb
@@ -39,6 +39,12 @@ def get_ios_denylist(workspace_root)
   denylist_config['ios'] || []
 end
 
+def is_version_at_least(current_version, minimum_version)
+  current_version ||= '0.0.0'
+  current_core = current_version.match(/^\d+(\.\d+)*/)&.to_s || '0.0.0'
+  Gem::Version.new(current_core) >= Gem::Version.new(minimum_version)
+end
+
 def rnrepo_pre_install(installer_context)
   # Check if plugin is disabled via environment variable
   if ENV['DISABLE_RNREPO']
@@ -69,7 +75,9 @@ def rnrepo_pre_install(installer_context)
 
   # Add worklets version to reanimated pod info
   if (pod = rn_pods.find { |p| p[:name] == 'RNReanimated' })
-    pod[:worklets_version] = worklets_version
+    if !is_version_at_least(pod[:version], "4.3.0")
+      pod[:worklets_version] = worklets_version
+    end
   end
 
   if rn_pods.empty?
@@ -321,13 +329,19 @@ module Pod
             end
           end
 
-          # Empty source files - CocoaPods will handle headers from xcframework automatically
-          spec.attributes_hash["source_files"] = []
+          # Create dummy header file
+          [debug_cache_dir, release_cache_dir, current_link].select { |d| Dir.exist?(d) }.each do |dir|
+            File.write(File.join(dir, 'dummy.h'), "// Dummy for #{pod_name}\n")
+          end
 
-          # Empty source files for each platform too
+          # Add dummy header as source files - so Xcode propagates info about the React Native framework to the main app target
+          dummy_header_path = ".rnrepo-cache/Current/dummy.h"
+          spec.attributes_hash["source_files"] = ["#{dummy_header_path}"]
+
+          # Add dummy header as source files for each platform too
           ["ios", "watchos", "tvos", "osx"].each do |platform|
             if spec.attributes_hash[platform]
-              spec.attributes_hash[platform]["source_files"] = []
+              spec.attributes_hash[platform]["source_files"] = ["#{dummy_header_path}"]
             end
           end