+++ packages/build-tools/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added support for ExpoModulesCore in Android release builds (#301)
+
+### Changed
+- worklets classifier is no longer added in reanimated@4.3.0+ artifacts (#317)
+
+### Fixed
+- added dummy header for iOS pods source_files (#311)
+
 ## [0.1.2-beta.0] - 2026-03-26
 
 ### Changed